+++ Runtime/TreasuredAction.cs
@@ -22,11 +22,12 @@ namespace Treasured.SDK
         [StringSelector("openLink", "showText", "media/PlayAudio", "media/PlayVideo", "selectObject")]
         private string _type;
         [SerializeField]
+        [TextArea]
         private string _src;
         [SerializeField]
         private string _targetId;
         [SerializeField]
-        private DisplayMode _displayMode;
+        private DisplayMode _displayMode = DisplayMode.Right;
         [SerializeField]
         [TextArea]
         private string _content;